Split LSP symbols.go into document, navigation, workspace, and call-hierarchy files
