+++ internal/lsp/symbols.go
@@ -2,17 +2,13 @@ package lsp
 
 import (
 	"bytes"
-	"encoding/json"
 	"net/url"
-	"path"
 	"path/filepath"
-	"sort"
 	"strings"
 
 	"github.com/jeduden/mdsmith/internal/config"
 	"github.com/jeduden/mdsmith/internal/discovery"
 	"github.com/jeduden/mdsmith/internal/index"
-	"github.com/jeduden/mdsmith/internal/linkgraph"
 	"github.com/jeduden/mdsmith/internal/lint"
 	"github.com/jeduden/mdsmith/internal/mdtext"
 	"github.com/jeduden/mdsmith/internal/yamlutil"
@@ -438,191 +434,6 @@ func isMarkdownExt(p string) bool {
 	return ext == ".md" || ext == ".markdown"
 }
 
-// handleDocumentSymbol returns a hierarchical outline of the buffer.
-// Front-matter keys hang off a synthetic top-of-file symbol;
-// directives become children of their enclosing heading.
-func (s *Server) handleDocumentSymbol(msg *requestMessage) {
-	var p documentSymbolParams
-	if err := json.Unmarshal(msg.Params, &p); err != nil {
-		_ = s.t.writeError(msg.ID, codeInvalidParams, "invalid documentSymbol params")
-		return
-	}
-	source, _, ok := s.docTextOrFile(p.TextDocument.URI)
-	if !ok {
-		_ = s.t.writeResponse(msg.ID, []documentSymbol{})
-		return
-	}
-	// Outline is built from the live source so unsaved edits are
-	// reflected even if the index hasn't been refreshed yet.
-	out := buildOutline(source)
-	_ = s.t.writeResponse(msg.ID, out)
-}
-
-// buildOutline turns a freshly parsed FileEntry into an LSP
-// hierarchical outline. Headings stack by level; front-matter keys
-// gather under a synthetic top-of-file node; directives attach to
-// their enclosing heading or to the file root.
-func buildOutline(source []byte) []documentSymbol {
-	idx := index.New("")
-	idx.Update("buffer", source)
-	fe, ok := idx.File("buffer")
-	if !ok {
-		return nil
-	}
-
-	var fmKids []documentSymbol
-	var dirRoot []documentSymbol
-	var headings []index.Symbol
-	for _, sym := range fe.Symbols {
-		switch sym.Kind {
-		case index.SymbolFrontMatter:
-			fmKids = append(fmKids, leafSymbol(sym, source))
-		case index.SymbolDirective:
-			dirRoot = append(dirRoot, leafSymbol(sym, source))
-		case index.SymbolHeading:
-			headings = append(headings, sym)
-		case index.SymbolLinkRef:
-			dirRoot = append(dirRoot, leafSymbol(sym, source))
-		}
-	}
-
-	var roots []documentSymbol
-	if len(fmKids) > 0 {
-		// Synthetic "front matter" parent at line 1.
-		roots = append(roots, documentSymbol{
-			Name:           "front matter",
-			Kind:           symbolKindProperty,
-			Range:          rangeForLines(1, 1, source),
-			SelectionRange: rangeForLines(1, 1, source),
-			Children:       fmKids,
-		})
-	}
-
-	hroots := buildHeadingTree(headings, source)
-	// Attach directives + link-refs whose line falls under a heading
-	// span; everything else hoists to the file root.
-	hroots, unattached := attachDirectives(hroots, dirRoot)
-	roots = append(roots, hroots...)
-	roots = append(roots, unattached...)
-	return roots
-}
-
-// buildHeadingTree turns a flat heading list into a nested
-// documentSymbol tree using a level-aware stack walk.
-func buildHeadingTree(headings []index.Symbol, source []byte) []documentSymbol {
-	var roots []documentSymbol
-	type stackEntry struct {
-		level int
-		node  *documentSymbol
-	}
-	var stack []stackEntry
-	for _, h := range headings {
-		ds := documentSymbol{
-			Name:           headingDisplay(h),
-			Detail:         headingDetail(h),
-			Kind:           symbolKindString,
-			Range:          rangeForLines(h.StartLine, h.EndLine, source),
-			SelectionRange: rangeForLines(h.SelectionLine, h.SelectionLine, source),
-		}
-		// Pop until we find a parent with a lower level.
-		for len(stack) > 0 && stack[len(stack)-1].level >= h.Level {
-			stack = stack[:len(stack)-1]
-		}
-		if len(stack) == 0 {
-			roots = append(roots, ds)
-			stack = append(stack, stackEntry{
-				level: h.Level,
-				node:  &roots[len(roots)-1],
-			})
-		} else {
-			parent := stack[len(stack)-1].node
-			parent.Children = append(parent.Children, ds)
-			stack = append(stack, stackEntry{
-				level: h.Level,
-				node:  &parent.Children[len(parent.Children)-1],
-			})
-		}
-	}
-	return roots
-}
-
-// attachDirectives walks the heading tree and reparents each
-// directive/leaf into the deepest heading whose range covers its
-// start line. Leaves that don't fall under any heading return as
-// the second value so the caller can hoist them to the file root.
-func attachDirectives(headings []documentSymbol, leaves []documentSymbol) ([]documentSymbol, []documentSymbol) {
-	var unattached []documentSymbol
-	for _, leaf := range leaves {
-		startLine := leaf.SelectionRange.Start.Line + 1 // back to 1-based
-		if !attachInto(headings, leaf, startLine) {
-			unattached = append(unattached, leaf)
-		}
-	}
-	return headings, unattached
-}
-
-func attachInto(nodes []documentSymbol, leaf documentSymbol, startLine int) bool {
-	for i := range nodes {
-		// LSP ranges are [start, end) in 0-based form. The leaf's
-		// start line lives inside the node when it falls between
-		// the node's Range start and end (inclusive).
-		nodeStart := nodes[i].Range.Start.Line + 1
-		nodeEnd := nodes[i].Range.End.Line + 1
-		if startLine >= nodeStart && startLine <= nodeEnd {
-			// Try to attach into a deeper child first.
-			if attachInto(nodes[i].Children, leaf, startLine) {
-				return true
-			}
-			nodes[i].Children = append(nodes[i].Children, leaf)
-			return true
-		}
-	}
-	return false
-}
-
-func leafSymbol(sym index.Symbol, source []byte) documentSymbol {
-	kind := symbolKindKey
-	switch sym.Kind {
-	case index.SymbolFrontMatter:
-		kind = symbolKindProperty
-	case index.SymbolDirective:
-		kind = symbolKindEvent
-	case index.SymbolLinkRef:
-		kind = symbolKindKey
-	}
-	return documentSymbol{
-		Name:           sym.Name,
-		Detail:         leafDetail(sym),
-		Kind:           kind,
-		Range:          rangeForLines(sym.StartLine, sym.EndLine, source),
-		SelectionRange: rangeForLines(sym.SelectionLine, sym.SelectionLine, source),
-	}
-}
-
-func headingDisplay(h index.Symbol) string {
-	if h.Name == "" {
-		return strings.Repeat("#", h.Level)
-	}
-	return h.Name
-}
-
-func headingDetail(h index.Symbol) string {
-	if h.Anchor == "" {
-		return ""
-	}
-	return "#" + h.Anchor
-}
-
-func leafDetail(sym index.Symbol) string {
-	switch sym.Kind {
-	case index.SymbolDirective:
-		return "<?" + sym.Name + "?>"
-	case index.SymbolLinkRef:
-		return "[" + sym.Name + "]:"
-	}
-	return ""
-}
-
 // rangeForLines returns an LSP Range covering 1-based start..end
 // lines inclusive. Columns are 0..end-of-line. Both bounds are
 // clamped to the document's line count so the emitted Range stays
@@ -695,700 +506,4 @@ func rangeAt(line, col int, source []byte) Range {
 		Start: Position{Line: line - 1, Character: startCh},
 		End:   Position{Line: line - 1, Character: endCh},
 	}
-}
-
-// handleDefinition resolves textDocument/definition.
-func (s *Server) handleDefinition(msg *requestMessage) {
-	var p textDocumentPositionParams
-	if err := json.Unmarshal(msg.Params, &p); err != nil {
-		_ = s.t.writeError(msg.ID, codeInvalidParams, "invalid definition params")
-		return
-	}
-	locs := s.resolveTargets(p, false)
-	if len(locs) == 0 {
-		_ = s.t.writeResponse(msg.ID, nil)
-		return
-	}
-	_ = s.t.writeResponse(msg.ID, locs[0])
-}
-
-// handleImplementation returns every match. For most tags this is the
-// same answer as Definition; only `kind:` values and headings (with
-// references) produce multi-target sets.
-func (s *Server) handleImplementation(msg *requestMessage) {
-	var p textDocumentPositionParams
-	if err := json.Unmarshal(msg.Params, &p); err != nil {
-		_ = s.t.writeError(msg.ID, codeInvalidParams, "invalid implementation params")
-		return
-	}
-	locs := s.resolveTargets(p, true)
-	_ = s.t.writeResponse(msg.ID, locs)
-}
-
-// resolveTargets is the shared core for definition and implementation.
-// When wantAll is false the slice is truncated to the first match.
-func (s *Server) resolveTargets(p textDocumentPositionParams, wantAll bool) []location {
-	source, rel, ok := s.docTextOrFile(p.TextDocument.URI)
-	if !ok {
-		return nil
-	}
-	idx := s.ensureIndex()
-
-	line := p.Position.Line + 1
-	col := lspPositionToByteColumn(source, line, p.Position.Character)
-	res := index.Locator{Path: rel}.Locate(source, line, col)
-	return s.resolveByTag(p, res, line, source, rel, idx, wantAll)
-}
-
-// resolveByTag dispatches on the locator's TokenTag and returns the
-// matching navigation targets. Split out of resolveTargets so the
-// switch stays small enough for funlen.
-func (s *Server) resolveByTag(
-	p textDocumentPositionParams, res index.LocateResult,
-	line int, source []byte, rel string, idx *index.Index, wantAll bool,
-) []location {
-	switch res.Tag {
-	case index.TokenAnchorLink:
-		return s.locationsForAnchor(rel, res.TargetAnchor, idx, source)
-	case index.TokenFileLink:
-		return s.locationsForFileLink(res.TargetFile, res.TargetAnchor, idx)
-	case index.TokenRefUse, index.TokenRefDef:
-		if loc, ok := s.locationForRefDef(rel, res.Label, source); ok {
-			return []location{loc}
-		}
-	case index.TokenDirectiveArg:
-		return s.directiveArgLocations(rel, res.DirectiveTargetFile)
-	case index.TokenHeading:
-		return s.headingTargets(p, rel, res.Anchor, line, source, idx, wantAll)
-	case index.TokenFileTop:
-		return []location{{
-			URI:   p.TextDocument.URI,
-			Range: rangeAt(1, 1, source),
-		}}
-	case index.TokenFrontMatterValue:
-		return s.frontMatterValueTargets(res.FrontMatterKey, res.FrontMatterValue, idx, wantAll)
-	}
-	return nil
-}
-
-func (s *Server) directiveArgLocations(rel, target string) []location {
-	if target == "" {
-		return nil
-	}
-	tgt := linkgraph.ResolveRelTarget(rel, target)
-	if tgt == "" {
-		return nil
-	}
-	return []location{{
-		URI:   s.workspaceURI(tgt),
-		Range: Range{Start: Position{Line: 0, Character: 0}, End: Position{Line: 0, Character: 0}},
-	}}
-}
-
-// headingTargets returns the heading itself for definition, plus
-// every link to it for implementation.
-func (s *Server) headingTargets(
-	p textDocumentPositionParams, rel, anchor string,
-	line int, source []byte, idx *index.Index, wantAll bool,
-) []location {
-	decl := []location{{
-		URI:   p.TextDocument.URI,
-		Range: rangeAt(line, 1, source),
-	}}
-	if !wantAll {
-		return decl
-	}
-	return append(decl, s.locationsForRefsToHeading(rel, anchor, idx)...)
-}
-
-// frontMatterValueTargets handles the `kind:` / `kinds:` value arm:
-// definition resolves to the kind block in `.mdsmith.yml`,
-// implementation widens to every file with that kind.
-func (s *Server) frontMatterValueTargets(key, val string, idx *index.Index, wantAll bool) []location {
-	if key != "kind" && key != "kinds" {
-		return nil
-	}
-	defs := s.locationsForKindDefinition(val)
-	if !wantAll {
-		return defs
-	}
-	return append(defs, s.locationsForFilesByKind(val, idx)...)
-}
-
-// locationsForAnchor returns the in-file heading targeted by an
-// anchor reference. It always returns at most one location — the
-// matching heading itself; multi-target widening for headings (the
-// implementation behavior) lives in resolveTargets' TokenHeading
-// arm, where the declaration is paired with all incoming links.
-func (s *Server) locationsForAnchor(rel, anchor string, idx *index.Index, source []byte) []location {
-	if anchor == "" {
-		return nil
-	}
-	if fe, ok := idx.File(rel); ok {
-		for _, sym := range fe.Symbols {
-			if sym.Kind == index.SymbolHeading && sym.Anchor == anchor {
-				return []location{{
-					URI:   s.workspaceURI(rel),
-					Range: rangeAt(sym.SelectionLine, sym.SelectionCol, source),
-				}}
-			}
-		}
-	}
-	return nil
-}
-
-// locationsForFileLink resolves `[text](./other.md#anchor)` to either
-// a heading in the target file or the file's first line.
-func (s *Server) locationsForFileLink(targetFile, anchor string, idx *index.Index) []location {
-	tgt := index.NormalizePath(targetFile)
-	if tgt == "" {
-		return nil
-	}
-	if anchor == "" {
-		return []location{{
-			URI:   s.workspaceURI(tgt),
-			Range: Range{Start: Position{Line: 0, Character: 0}, End: Position{Line: 0, Character: 0}},
-		}}
-	}
-	fe, ok := idx.File(tgt)
-	if !ok {
-		// File lives outside the index (or wasn't loaded yet).
-		// Return a best-effort target at line 1.
-		return []location{{
-			URI:   s.workspaceURI(tgt),
-			Range: Range{Start: Position{Line: 0, Character: 0}, End: Position{Line: 0, Character: 0}},
-		}}
-	}
-	for _, sym := range fe.Symbols {
-		if sym.Kind == index.SymbolHeading && sym.Anchor == anchor {
-			return []location{{
-				URI:   s.workspaceURI(tgt),
-				Range: rangeAt(sym.SelectionLine, sym.SelectionCol, nil),
-			}}
-		}
-	}
-	return []location{{
-		URI:   s.workspaceURI(tgt),
-		Range: Range{Start: Position{Line: 0, Character: 0}, End: Position{Line: 0, Character: 0}},
-	}}
-}
-
-// locationForRefDef returns the position of `[label]: …` in the
-// current file.
-func (s *Server) locationForRefDef(rel, label string, source []byte) (location, bool) {
-	idx := s.ensureIndex()
-	fe, ok := idx.File(rel)
-	if !ok {
-		return location{}, false
-	}
-	for _, sym := range fe.Symbols {
-		if sym.Kind == index.SymbolLinkRef && sym.Anchor == label {
-			return location{
-				URI:   s.workspaceURI(rel),
-				Range: rangeAt(sym.SelectionLine, sym.SelectionCol, source),
-			}, true
-		}
-	}
-	return location{}, false
-}
-
-// locationsForRefsToHeading scans every file's outgoing edges for
-// references to (rel, anchor) and returns one location per match.
-func (s *Server) locationsForRefsToHeading(rel, anchor string, idx *index.Index) []location {
-	if anchor == "" {
-		return nil
-	}
-	edges := idx.IncomingEdges(rel, anchor)
-	out := make([]location, 0, len(edges))
-	for _, e := range edges {
-		out = append(out, location{
-			URI:   s.workspaceURI(e.SourceFile),
-			Range: rangeAt(e.SourceLine, e.SourceCol, nil),
-		})
-	}
-	sort.Slice(out, func(i, j int) bool {
-		if out[i].URI != out[j].URI {
-			return out[i].URI < out[j].URI
-		}
-		return out[i].Range.Start.Line < out[j].Range.Start.Line
-	})
-	return out
-}
-
-// locationsForKindDefinition reports the location of the kind block
-// in `.mdsmith.yml`. We surface the config file at line 1 when the
-// kind is declared; absent kinds yield nothing.
-func (s *Server) locationsForKindDefinition(kind string) []location {
-	cfg, configPath, _ := s.snapshotConfig()
-	if cfg == nil || configPath == "" {
-		return nil
-	}
-	if _, ok := cfg.Kinds[kind]; !ok {
-		return nil
-	}
-	return []location{{
-		URI:   pathToURI(configPath),
-		Range: Range{Start: Position{Line: 0, Character: 0}, End: Position{Line: 0, Character: 0}},
-	}}
-}
-
-// locationsForFilesByKind returns one Location per workspace file
-// whose front-matter `kinds:` includes kind.
-func (s *Server) locationsForFilesByKind(kind string, idx *index.Index) []location {
-	files := idx.FilesByKind(kind)
-	out := make([]location, 0, len(files))
-	for _, rel := range files {
-		out = append(out, location{
-			URI:   s.workspaceURI(rel),
-			Range: Range{Start: Position{Line: 0, Character: 0}, End: Position{Line: 0, Character: 0}},
-		})
-	}
-	sort.Slice(out, func(i, j int) bool { return out[i].URI < out[j].URI })
-	return out
-}
-
-// handleReferences resolves textDocument/references.
-func (s *Server) handleReferences(msg *requestMessage) {
-	var p referencesParams
-	if err := json.Unmarshal(msg.Params, &p); err != nil {
-		_ = s.t.writeError(msg.ID, codeInvalidParams, "invalid references params")
-		return
-	}
-	source, rel, ok := s.docTextOrFile(p.TextDocument.URI)
-	if !ok {
-		_ = s.t.writeResponse(msg.ID, []location{})
-		return
-	}
-	idx := s.ensureIndex()
-	line := p.Position.Line + 1
-	col := lspPositionToByteColumn(source, line, p.Position.Character)
-	res := index.Locator{Path: rel}.Locate(source, line, col)
-
-	var out []location
-	switch res.Tag {
-	case index.TokenHeading:
-		out = s.locationsForRefsToHeading(rel, res.Anchor, idx)
-		if p.Context.IncludeDeclaration {
-			out = prependLocation(out, location{
-				URI:   p.TextDocument.URI,
-				Range: rangeAt(p.Position.Line+1, 1, source),
-			})
-		}
-	case index.TokenRefDef:
-		// Every reference-style use of `label` in this file.
-		out = s.locationsForRefUses(rel, res.Label, idx)
-		if p.Context.IncludeDeclaration {
-			if loc, ok := s.locationForRefDef(rel, res.Label, source); ok {
-				out = prependLocation(out, loc)
-			}
-		}
-	case index.TokenFileTop:
-		// Every link target that names this file with no anchor.
-		out = s.locationsForFileTop(rel, idx)
-	case index.TokenFrontMatterValue:
-		if res.FrontMatterKey == "kind" || res.FrontMatterKey == "kinds" {
-			out = s.locationsForFilesByKind(res.FrontMatterValue, idx)
-		}
-	case index.TokenDirectiveArg:
-		// References on a directive argument resolve to "every
-		// workspace edge that points at this file" — file links
-		// (no anchor) plus every <?include?>, <?build?>, and
-		// <?catalog?>. Limiting to EdgeFileLink (the previous
-		// behavior) hid the directive-to-directive references that
-		// users actually need when navigating include / build chains.
-		if res.DirectiveTargetFile != "" {
-			if tgt := linkgraph.ResolveRelTarget(rel, res.DirectiveTargetFile); tgt != "" {
-				out = s.locationsForFileReferences(tgt, idx)
-			}
-		}
-	}
-	if out == nil {
-		out = []location{}
-	}
-	_ = s.t.writeResponse(msg.ID, out)
-}
-
-func prependLocation(rest []location, loc location) []location {
-	out := make([]location, 0, len(rest)+1)
-	out = append(out, loc)
-	out = append(out, rest...)
-	return out
-}
-
-// locationsForRefUses returns every `[text][label]` in rel.
-func (s *Server) locationsForRefUses(rel, label string, idx *index.Index) []location {
-	fe, ok := idx.File(rel)
-	if !ok {
-		return nil
-	}
-	var out []location
-	for _, e := range fe.Outgoing {
-		if e.Kind != index.EdgeRefLink {
-			continue
-		}
-		if !strings.EqualFold(e.TargetLabel, label) {
-			continue
-		}
-		out = append(out, location{
-			URI:   s.workspaceURI(rel),
-			Range: rangeAt(e.SourceLine, e.SourceCol, nil),
-		})
-	}
-	return out
-}
-
-// locationsForFileTop returns every workspace link whose path
-// component points at file (with empty anchor).
-func (s *Server) locationsForFileTop(file string, idx *index.Index) []location {
-	edges := idx.IncomingEdges(file, "")
-	var out []location
-	for _, e := range edges {
-		if e.Kind != index.EdgeFileLink {
-			continue
-		}
-		if e.TargetAnchor != "" {
-			continue
-		}
-		out = append(out, location{
-			URI:   s.workspaceURI(e.SourceFile),
-			Range: rangeAt(e.SourceLine, e.SourceCol, nil),
-		})
-	}
-	sort.Slice(out, func(i, j int) bool {
-		if out[i].URI != out[j].URI {
-			return out[i].URI < out[j].URI
-		}
-		return out[i].Range.Start.Line < out[j].Range.Start.Line
-	})
-	return out
-}
-
-// locationsForFileReferences returns every workspace edge whose
-// target is file: the union of file-top links and the include /
-// build / catalog directives that target this file. Reference-style
-// link uses are not included because they target a label, not a
-// file path.
-func (s *Server) locationsForFileReferences(file string, idx *index.Index) []location {
-	edges := idx.IncomingEdges(file, "")
-	var out []location
-	for _, e := range edges {
-		switch e.Kind {
-		case index.EdgeFileLink:
-			if e.TargetAnchor != "" {
-				continue
-			}
-		case index.EdgeInclude, index.EdgeBuild, index.EdgeCatalog:
-			// keep
-		default:
-			continue
-		}
-		out = append(out, location{
-			URI:   s.workspaceURI(e.SourceFile),
-			Range: rangeAt(e.SourceLine, e.SourceCol, nil),
-		})
-	}
-	sort.Slice(out, func(i, j int) bool {
-		if out[i].URI != out[j].URI {
-			return out[i].URI < out[j].URI
-		}
-		return out[i].Range.Start.Line < out[j].Range.Start.Line
-	})
-	return out
-}
-
-// handleWorkspaceSymbol returns SymbolInformation entries for every
-// substring match in the workspace index.
-func (s *Server) handleWorkspaceSymbol(msg *requestMessage) {
-	var p workspaceSymbolParams
-	if err := json.Unmarshal(msg.Params, &p); err != nil {
-		_ = s.t.writeError(msg.ID, codeInvalidParams, "invalid workspace/symbol params")
-		return
-	}
-	idx := s.ensureIndex()
-	hits := idx.SearchSymbols(p.Query, 1024)
-	out := make([]symbolInformation, 0, len(hits))
-	for _, h := range hits {
-		kind := symbolKindString
-		switch h.Symbol.Kind {
-		case index.SymbolFrontMatter:
-			kind = symbolKindProperty
-		case index.SymbolLinkRef:
-			kind = symbolKindKey
-		case index.SymbolDirective:
-			kind = symbolKindEvent
-		}
-		out = append(out, symbolInformation{
-			Name: h.Symbol.Name,
-			Kind: kind,
-			Location: location{
-				URI:   s.workspaceURI(h.File),
-				Range: rangeAt(h.Symbol.SelectionLine, h.Symbol.SelectionCol, nil),
-			},
-			ContainerName: h.File,
-		})
-	}
-	sort.Slice(out, func(i, j int) bool {
-		if out[i].ContainerName != out[j].ContainerName {
-			return out[i].ContainerName < out[j].ContainerName
-		}
-		return out[i].Name < out[j].Name
-	})
-	_ = s.t.writeResponse(msg.ID, out)
-}
-
-// handlePrepareCallHierarchy returns a single call-hierarchy item
-// anchored at (file, optional heading). On a directive arg the item
-// is the target file; on a heading line, the heading section.
-func (s *Server) handlePrepareCallHierarchy(msg *requestMessage) {
-	var p textDocumentPositionParams
-	if err := json.Unmarshal(msg.Params, &p); err != nil {
-		_ = s.t.writeError(msg.ID, codeInvalidParams, "invalid prepareCallHierarchy params")
-		return
-	}
-	source, rel, ok := s.docTextOrFile(p.TextDocument.URI)
-	if !ok {
-		_ = s.t.writeResponse(msg.ID, []callHierarchyItem{})
-		return
-	}
-	idx := s.ensureIndex()
-	line := p.Position.Line + 1
-	col := lspPositionToByteColumn(source, line, p.Position.Character)
-	res := index.Locator{Path: rel}.Locate(source, line, col)
-
-	var item callHierarchyItem
-	switch res.Tag {
-	case index.TokenHeading:
-		fe, _ := idx.File(rel)
-		item = callHierarchyItem{
-			Name:           res.Name,
-			Kind:           symbolKindString,
-			Detail:         "#" + res.Anchor,
-			URI:            p.TextDocument.URI,
-			Range:          headingRangeFromIndex(rel, res.Anchor, fe, source),
-			SelectionRange: rangeAt(p.Position.Line+1, 1, source),
-			Data:           &callHierarchyData{File: rel, Anchor: res.Anchor},
-		}
-	case index.TokenDirectiveArg:
-		if res.DirectiveTargetFile != "" {
-			tgt := linkgraph.ResolveRelTarget(rel, res.DirectiveTargetFile)
-			if tgt == "" {
-				_ = s.t.writeResponse(msg.ID, []callHierarchyItem{})
-				return
-			}
-			item = callHierarchyItem{
-				Name:           tgt,
-				Kind:           symbolKindString,
-				URI:            s.workspaceURI(tgt),
-				Range:          Range{Start: Position{Line: 0, Character: 0}, End: Position{Line: 0, Character: 0}},
-				SelectionRange: Range{Start: Position{Line: 0, Character: 0}, End: Position{Line: 0, Character: 0}},
-				Data:           &callHierarchyData{File: tgt},
-			}
-		}
-	case index.TokenFileTop:
-		// File-level call hierarchy: only the very top of the
-		// document anchors at the file. Plain prose lower in the
-		// file does not — see Plan 131's cursor matrix
-		// (file root / heading / directive arg). Without this
-		// gate the editor would offer a "call hierarchy" entry
-		// for arbitrary positions, including paragraphs that
-		// have no inbound or outbound references.
-		item = callHierarchyItem{
-			Name:           rel,
-			Kind:           symbolKindString,
-			URI:            p.TextDocument.URI,
-			Range:          rangeForLines(1, lineCount(source), source),
-			SelectionRange: rangeAt(1, 1, source),
-			Data:           &callHierarchyData{File: rel},
-		}
-	}
-	if item.URI == "" {
-		_ = s.t.writeResponse(msg.ID, []callHierarchyItem{})
-		return
-	}
-	_ = s.t.writeResponse(msg.ID, []callHierarchyItem{item})
-}
-
-func headingRangeFromIndex(rel, anchor string, fe *index.FileEntry, source []byte) Range {
-	if fe == nil {
-		return rangeAt(1, 1, source)
-	}
-	for _, sym := range fe.Symbols {
-		if sym.Kind == index.SymbolHeading && sym.Anchor == anchor {
-			return rangeForLines(sym.StartLine, sym.EndLine, source)
-		}
-	}
-	return rangeAt(1, 1, source)
-}
-
-func lineCount(source []byte) int {
-	if len(source) == 0 {
-		return 1
-	}
-	n := bytes.Count(source, []byte{'\n'})
-	if source[len(source)-1] != '\n' {
-		n++
-	}
-	return n
-}
-
-// handleIncomingCalls returns every workspace edge into the item.
-// Edges from the same source file are coalesced into one entry with
-// multiple `fromRanges`; LSP clients render each fromRange as a
-// click target under the same caller, so emitting one item per edge
-// would show the same caller N times in the call-hierarchy view.
-func (s *Server) handleIncomingCalls(msg *requestMessage) {
-	var p callHierarchyIncomingCallsParams
-	if err := json.Unmarshal(msg.Params, &p); err != nil {
-		_ = s.t.writeError(msg.ID, codeInvalidParams, "invalid incomingCalls params")
-		return
-	}
-	if p.Item.Data == nil {
-		_ = s.t.writeResponse(msg.ID, []callHierarchyIncomingCall{})
-		return
-	}
-	idx := s.ensureIndex()
-	edges := idx.IncomingEdges(p.Item.Data.File, p.Item.Data.Anchor)
-
-	type bucket struct {
-		item   callHierarchyItem
-		ranges []Range
-	}
-	order := make([]string, 0, len(edges))
-	groups := make(map[string]*bucket, len(edges))
-	for _, e := range edges {
-		// Call hierarchy is a cross-file dependency view: keep only
-		// the edge kinds that represent inter-document flow. Anchor
-		// and reference-style links are intra-document, and an
-		// edge whose SourceFile equals the item's File is a
-		// self-reference (e.g. `[a](#sec)` to a heading whose
-		// anchor matches `Anchor`); both would clutter the result.
-		if e.Kind == index.EdgeAnchorLink || e.Kind == index.EdgeRefLink {
-			continue
-		}
-		if e.SourceFile == p.Item.Data.File {
-			continue
-		}
-		r := rangeAt(e.SourceLine, e.SourceCol, nil)
-		if g, ok := groups[e.SourceFile]; ok {
-			g.ranges = append(g.ranges, r)
-			continue
-		}
-		groups[e.SourceFile] = &bucket{
-			item: callHierarchyItem{
-				Name:           e.SourceFile,
-				Kind:           symbolKindString,
-				URI:            s.workspaceURI(e.SourceFile),
-				Range:          r,
-				SelectionRange: r,
-				Data:           &callHierarchyData{File: e.SourceFile},
-			},
-			ranges: []Range{r},
-		}
-		order = append(order, e.SourceFile)
-	}
-	out := make([]callHierarchyIncomingCall, 0, len(order))
-	for _, k := range order {
-		g := groups[k]
-		out = append(out, callHierarchyIncomingCall{From: g.item, FromRanges: g.ranges})
-	}
-	_ = s.t.writeResponse(msg.ID, out)
-}
-
-// handleOutgoingCalls returns every edge out of the item, scoped
-// to the section when the item carries an anchor (heading-level
-// call hierarchy). Edges to the same target file are coalesced into
-// one entry with multiple `fromRanges`, matching the LSP grouping
-// contract.
-func (s *Server) handleOutgoingCalls(msg *requestMessage) {
-	var p callHierarchyOutgoingCallsParams
-	if err := json.Unmarshal(msg.Params, &p); err != nil {
-		_ = s.t.writeError(msg.ID, codeInvalidParams, "invalid outgoingCalls params")
-		return
-	}
-	if p.Item.Data == nil {
-		_ = s.t.writeResponse(msg.ID, []callHierarchyOutgoingCall{})
-		return
-	}
-	idx := s.ensureIndex()
-	edges := idx.OutgoingEdges(p.Item.Data.File)
-	startLine, endLine := outgoingScope(idx, p.Item.Data)
-
-	type bucket struct {
-		item   callHierarchyItem
-		ranges []Range
-	}
-	order := make([]string, 0, len(edges))
-	groups := make(map[string]*bucket, len(edges))
-	for _, e := range edges {
-		// Same-file anchor / ref-style links are intra-document and
-		// don't fit the cross-file call-graph view.
-		if e.Kind == index.EdgeAnchorLink || e.Kind == index.EdgeRefLink {
-			continue
-		}
-		// Heading-scoped item: skip edges outside the section's
-		// source range so a heading with no outbound links doesn't
-		// inherit calls from sibling sections.
-		if endLine > 0 && (e.SourceLine < startLine || e.SourceLine > endLine) {
-			continue
-		}
-		toFile := e.TargetFile
-		if toFile == "" {
-			// Catalog without expansion: point at the host file's
-			// directory as a placeholder. Plan 131 documents this
-			// fallback explicitly under "Open Questions".
-			toFile = path.Dir(p.Item.Data.File)
-		}
-		r := rangeAt(e.SourceLine, e.SourceCol, nil)
-		if g, ok := groups[toFile]; ok {
-			g.ranges = append(g.ranges, r)
-			continue
-		}
-		// Coalesce by target file. The bucket represents the
-		// callee file as a whole, so Data.Anchor must stay empty
-		// — different edges from the source can target different
-		// headings inside the same file, and a follow-up
-		// incomingCalls on this item would otherwise be filtered
-		// to whichever anchor happened to land in the bucket
-		// first. To navigate to a specific heading, the user can
-		// open the callee and re-issue prepareCallHierarchy
-		// there.
-		groups[toFile] = &bucket{
-			item: callHierarchyItem{
-				Name:           toFile,
-				Kind:           symbolKindString,
-				URI:            s.workspaceURI(toFile),
-				Range:          Range{Start: Position{Line: 0, Character: 0}, End: Position{Line: 0, Character: 0}},
-				SelectionRange: Range{Start: Position{Line: 0, Character: 0}, End: Position{Line: 0, Character: 0}},
-				Data:           &callHierarchyData{File: toFile},
-			},
-			ranges: []Range{r},
-		}
-		order = append(order, toFile)
-	}
-	out := make([]callHierarchyOutgoingCall, 0, len(order))
-	for _, k := range order {
-		g := groups[k]
-		out = append(out, callHierarchyOutgoingCall{To: g.item, FromRanges: g.ranges})
-	}
-	_ = s.t.writeResponse(msg.ID, out)
-}
-
-// outgoingScope returns the [startLine, endLine] bound for outgoing
-// edges when the call-hierarchy item is heading-scoped. Returns
-// (1, 0) — i.e. an open-ended range — for file-level items so the
-// caller treats every edge as in scope.
-func outgoingScope(idx *index.Index, data *callHierarchyData) (int, int) {
-	if data == nil || data.Anchor == "" {
-		return 1, 0
-	}
-	fe, ok := idx.File(data.File)
-	if !ok {
-		return 1, 0
-	}
-	for _, sym := range fe.Symbols {
-		if sym.Kind == index.SymbolHeading && sym.Anchor == data.Anchor {
-			return sym.StartLine, sym.EndLine
-		}
-	}
-	return 1, 0
 }
